+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.1.0 (Unreleased)
 
 FEATURES:
+
+* **New Data Source:** `gigahost_account`